Complete TODOs
