+++ package.json
@@ -33,6 +33,7 @@
     "@tailwindcss/typography": "^0.5.16",
     "@types/hast": "^3.0.4",
     "@types/unist": "^3.0.3",
+    "@vcarl/remark-headings": "^0.1.0",
     "@vercel/analytics": "^1.4.1",
     "@vercel/speed-insights": "^1.1.0",
     "@yeskunall/astro-umami": "^0.0.3",
