+++ src/components/PostCard.astro
@@ -6,7 +6,7 @@ import { Icon } from 'astro-icon/components'
 import { getDir } from '../utils/url-utils'
 import I18nKey from '@i18n/i18nKey'
 import { i18n } from '@i18n/translation'
-import { type ReadingTime } from '@utils/parser'
+import type { ReadingTime } from '@utils/content-utils'
 
 interface Props {
   class?: string
