+++ src/components/PostPage.astro
@@ -1,7 +1,7 @@
 ---
 import { getPostUrlBySlug } from '@utils/url-utils'
 import PostCard from './PostCard.astro'
-import type { ReadingTime } from '@utils/parser'
+import type { ReadingTime } from '@utils/content-utils'
 
 const { page } = Astro.props
