+++ src/components/widget/Categories.astro
@@ -1,12 +1,11 @@
 ---
 import WidgetLayout from './WidgetLayout.astro'
 
-import { i18n } from '../../i18n/translation'
-import I18nKey from '../../i18n/i18nKey'
-//import { getCategoryList } from '../../utils/content-utils'
-import { getCategoryUrl } from '../../utils/url-utils'
-import ButtonLink from '../control/ButtonLink.astro'
-import { getCategoryList } from '@utils/parser'
+import { i18n } from '@i18n/translation'
+import I18nKey from '@i18n/i18nKey'
+import { getCategoryUrl } from '@utils/url-utils'
+import ButtonLink from '@components/control/ButtonLink.astro'
+import { getCategoryList } from '@utils/content-utils'
 
 const categories = await getCategoryList()
