+++ src/components/widget/SideBar.astro
@@ -3,7 +3,6 @@ import Profile from './Profile.astro'
 import Tag from './Tags.astro'
 import Categories from './Categories.astro'
 import type { MarkdownHeading } from 'astro'
-import TOC from './TOC.astro'
 
 interface Props {
     class? : string
