+++ src/components/widget/Tags.astro
@@ -2,11 +2,10 @@
 
 import WidgetLayout from './WidgetLayout.astro'
 import ButtonTag from '../control/ButtonTag.astro'
-//import { getTagList } from '../../utils/content-utils'
 import { i18n } from '../../i18n/translation'
 import I18nKey from '../../i18n/i18nKey'
 import { url } from '../../utils/url-utils'
-import { getAllTags } from '@utils/parser'
+import { getAllTags } from '@utils/content-utils'
 
 const tags = await getAllTags()
