+++ src/pages/archive/category/[category].astro
@@ -4,7 +4,7 @@ import MainGridLayout from '@layouts/MainGridLayout.astro'
 import ArchivePanel from '@components/ArchivePanel.astro'
 import { i18n } from '@i18n/translation'
 import I18nKey from '@i18n/i18nKey'
-import { getCategoryList } from '@utils/parser'
+import { getCategoryList } from '@utils/content-utils'
 
 export async function getStaticPaths() {
   const categories = await getCategoryList()
