+++ src/pages/archive/tag/[tag].astro
@@ -3,8 +3,7 @@ import ArchivePanel from '@components/ArchivePanel.astro'
 import I18nKey from '@i18n/i18nKey'
 import { i18n } from '@i18n/translation'
 import MainGridLayout from '@layouts/MainGridLayout.astro'
-import { getSortedPosts } from '@utils/parser'
-//import { getSortedPosts } from '@utils/content-utils'
+import { getSortedPosts } from '@utils/content-utils'
 
 export async function getStaticPaths() {
   const posts = await getSortedPosts()
