+++ src/pages/rss.xml.ts
@@ -1,8 +1,7 @@
 import { siteConfig } from '@/config'
 import rss from '@astrojs/rss'
-import { getSortedPosts } from '@utils/parser';
 import type { APIContext } from 'astro';
-//import { getSortedPosts } from '@utils/content-utils'
+import { getSortedPosts } from '@utils/content-utils'
 import MarkdownIt from 'markdown-it'
 import sanitizeHtml from 'sanitize-html'
 
@@ -23,18 +22,18 @@ export async function GET(context: APIContext) {
     items: blog.map(post => {
       return {
         title: post.data?.title,
-        pubDate: post.data?.published,
+        pubDate: new Date(post.data?.published as string) ?? new Date(),
         description: post.data?.description || '',
         link: `/posts/${post.slug}/`,
-        content: sanitizeHtml(parser.render(post.body), {
+        content: sanitizeHtml(parser.render(post.body as string), {
           allowedTags: sanitizeHtml.defaults.allowedTags.concat(['img']),
         }),
       }
     }),
     customData:
       `<language>${siteConfig.lang}</language>` +
-      `<lastBuildDate>${new Date(blog[0]?.data?.published).toUTCString()}</lastBuildDate>` +
+      `<lastBuildDate>${new Date(blog[0]?.data?.published as string).toUTCString()}</lastBuildDate>` +
       `<atom:link href="${context.site}rss.xml" rel="self" type="application/rss+xml"/>` +
-      `<pubDate>${new Date(blog[0]?.data.published).toUTCString()}</pubDate>`,
+      `<pubDate>${new Date(blog[0]?.data?.published as string).toUTCString()}</pubDate>`,
   })
 }
